+++ src/virtual-machine.cpp
@@ -36,6 +36,10 @@ interpreter_break_reason virtual_machine::do_run(uint64_t mcycle_end) {
     return m_machine->run(mcycle_end);
 }
 
+interpreter_break_reason virtual_machine::do_log_step(uint64_t mcycle_count, const std::string &filename) {
+    return m_machine->log_step(mcycle_count, filename);
+}
+
 access_log virtual_machine::do_log_uarch_step(const access_log::type &log_type, bool one_based) {
     return m_machine->log_uarch_step(log_type, one_based);
 }